Move example sketch to an appropriately named subfolder
The Arduino IDE requires sketches to be located in a folder that matches the filename. Previously, the example sketch did not appear under the File > Examples menu and attempting to open the sketch file would cause the Arduino IDE to display the message:

The file "sleepy_blink.ino" needs to be inside a sketch folder named "sleepy_blink". Create this folder, move file, and continue?